+++ src/spawn/builtin/mod.rs
@@ -1,5 +1,7 @@
 //! Defines methods for spawning shell builtin commands
 
+mod colon;
 mod shift;
 
+pub use self::colon::{Colon, colon, SpawnedColon};
 pub use self::shift::{Shift, shift, SpawnedShift};