+++ CHANGELOG.md
@@ -12,6 +12,7 @@ backwards compatible manner.
 - Added `ShiftArgumentsEnvironment` as an interface for shifting positional arguments.
 - Added a `spawn::builtin` module for hosting shell-builtin command implementations
 - Added a builtin implementation for the `shift` shell-command
+- Added a builtin implementation for the `:` shell-command
 
 ### Changed
 - Reduced required bounds for implementing `VarEnvRestorer` to just `E: VariableEnvironment`